// src/lib/api/generated/model/businessDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Agendly.Api
 * OpenAPI spec version: 1.0
 */

export interface BusinessDto {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  active: boolean;
  timeZoneId: string;
}